+++ src/lib/Scenes/City/City.tsx
@@ -1,15 +1,16 @@
-import { color, Flex, Theme } from "@artsy/palette"
+import { Box, color, Flex, Sans, Theme } from "@artsy/palette"
 import ScrollableTabBar, { ScrollableTab } from "lib/Components/ScrollableTabBar"
 import { Schema, screenTrack } from "lib/utils/track"
 import React, { Component } from "react"
 import ScrollableTabView from "react-native-scrollable-tab-view"
 
+import InvertedButton from "lib/Components/Buttons/InvertedButton"
 import { NativeModules, View } from "react-native"
 import { RelayProp } from "react-relay"
 import styled from "styled-components/native"
 import { BucketResults } from "../Map/bucketCityResults"
 import { EventEmitter } from "../Map/EventEmitter"
-import { MapTab } from "../Map/types"
+import { MapTab, RelayErrorState } from "../Map/types"
 import { cityTabs } from "./cityTabs"
 import { AllEvents } from "./Components/AllEvents"
 import { EventList } from "./Components/EventList"
@@ -30,6 +31,7 @@ interface State {
   citySlug: string
   selectedTab: number
   sponsoredContent: { introText: string; artGuideUrl: string }
+  relayErrorState?: RelayErrorState
 }
 const AllCityMetaTab = 0
 
@@ -67,6 +69,7 @@ export class CityView extends Component<Props, State> {
     selectedTab: AllCityMetaTab,
     citySlug: "",
     sponsoredContent: null,
+    relayErrorState: null,
   }
 
   scrollViewVerticalStart = 0
@@ -86,13 +89,26 @@ export class CityView extends Component<Props, State> {
     citySlug: string
     sponsoredContent: { introText: string; artGuideUrl: string }
   }) => {
-    this.setState({
-      buckets,
-      filter,
-      cityName,
-      citySlug,
-      relay,
-      sponsoredContent,
+    // We have the Relay response; post a notification so that the ARMapContainerViewController can finalize the native UI (ie: show the drawer partially).
+    this.setState(
+      {
+        buckets,
+        filter,
+        cityName,
+        citySlug,
+        relay,
+        sponsoredContent,
+      },
+      () => {
+        NativeModules.ARNotificationsManager.postNotificationName("ARLocalDiscoveryQueryResponseReceived", {})
+      }
+    )
+  }
+
+  handleError = ({ relayErrorState }: { relayErrorState: RelayErrorState }) => {
+    // We have a Relay error; post a notification so that the ARMapContainerViewController can finalize the native UI (ie: show the drawer partially).
+    this.setState({ relayErrorState }, () => {
+      NativeModules.ARNotificationsManager.postNotificationName("ARLocalDiscoveryQueryResponseReceived", {})
     })
   }
 
@@ -112,17 +128,17 @@ export class CityView extends Component<Props, State> {
 
   componentWillMount() {
     EventEmitter.subscribe("map:change", this.handleEvent)
+    EventEmitter.subscribe("map:error", this.handleError)
   }
 
   componentWillUnmount() {
     EventEmitter.unsubscribe("map:change", this.handleEvent)
-    // this.trackSession()
+    EventEmitter.unsubscribe("map:error", this.handleError)
   }
 
-  componentDidUpdate() {
-    if (this.state.buckets) {
-      // We have the Relay response; post a notification so that the ARMapContainerViewController can finalize the native UI.
-      NativeModules.ARNotificationsManager.postNotificationName("ARLocalDiscoveryQueryResponseReceived", {})
+  componentWillReceiveProps(nextProps) {
+    if (this.props.isDrawerOpen !== nextProps.isDrawerOpen) {
+      this.fireScreenViewAnalytics()
     }
   }
 
@@ -145,19 +161,21 @@ export class CityView extends Component<Props, State> {
   }
 
   render() {
-    const { buckets, cityName, citySlug } = this.state
+    const { buckets, cityName, citySlug, relayErrorState } = this.state
     const { verticalMargin } = this.props
     // bottomInset is used for the ScrollView's contentInset. See the note in ARMapContainerViewController.m for context.
     const bottomInset = this.scrollViewVerticalStart + (verticalMargin || 0)
 
-    return (
-      buckets && (
-        <Theme>
-          <>
-            <Flex style={{ flex: 1 }}>
-              <Flex py={1} alignItems="center">
-                <Handle />
-              </Flex>
+    return buckets || relayErrorState ? (
+      <Theme>
+        <>
+          <Flex style={{ flex: 1 }}>
+            <Flex py={1} alignItems="center">
+              <Handle />
+            </Flex>
+            {relayErrorState ? (
+              <ErrorScreen relayErrorState={relayErrorState} key="error" />
+            ) : (
               <ScrollableTabView
                 initialPage={this.props.initialTab || AllCityMetaTab}
                 onChangeTab={selectedTab => this.setSelectedTab(selectedTab)}
@@ -202,11 +220,11 @@ export class CityView extends Component<Props, State> {
                   )
                 })}
               </ScrollableTabView>
-            </Flex>
-          </>
-        </Theme>
-      )
-    )
+            )}
+          </Flex>
+        </>
+      </Theme>
+    ) : null
   }
 }
 
@@ -216,3 +234,18 @@ const Handle = styled.View`
   border-radius: 2.5px;
   background-color: ${color("black30")};
 `
+
+const ErrorScreen: React.SFC<{ relayErrorState: RelayErrorState }> = ({ relayErrorState: { retry, isRetrying } }) => {
+  return (
+    <Box py={2}>
+      <Sans size="3t" textAlign="center" mx={2}>
+        We are having trouble loading content right now, please try again later.
+      </Sans>
+      <Flex justifyContent="center" flexDirection="row">
+        <Box width={69} height={34} mt={2}>
+          <InvertedButton text="Retry" onPress={retry} inProgress={isRetrying} />
+        </Box>
+      </Flex>
+    </Box>
+  )
+}