+++ src/lib/Scenes/Map/EventEmitter.ts
@@ -1,6 +1,6 @@
 type Callback = (data: any) => void
 
-type Event = "map:change" | "filters:change"
+type Event = "map:change" | "map:error" | "filters:change"
 
 // tslint:disable-next-line:class-name
 class _EventEmitter {