+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Master
 
+- Adds retry screen to map view - ash
 - Fixes no image state for artist lists - kieran
 - Adds Mapbox logo and attribution for licensing - luc
 - Fix bug with selected pin and clusters not deselecting - luc